+++ crates/editor/src/signature_help.rs
@@ -1,19 +1,19 @@
-mod popover;
-mod state;
-
 use crate::actions::ShowSignatureHelp;
 use crate::{Editor, EditorSettings, ToggleAutoSignatureHelp};
-use gpui::{combine_highlights, App, Context, TextStyle, Window};
+use gpui::{
+    combine_highlights, App, Context, HighlightStyle, MouseButton, Size, StyledText, Task,
+    TextStyle, Window,
+};
 use language::BufferSnapshot;
 use multi_buffer::{Anchor, ToOffset};
 use settings::Settings;
 use std::ops::Range;
 use text::Rope;
 use theme::ThemeSettings;
-use ui::{relative, ActiveTheme};
-
-pub use popover::SignatureHelpPopover;
-pub use state::SignatureHelpState;
+use ui::{
+    div, relative, ActiveTheme, AnyElement, InteractiveElement, IntoElement, ParentElement, Pixels,
+    SharedString, StatefulInteractiveElement, Styled, StyledExt,
+};
 
 // Language-specific settings may define quotes as "brackets", so filter them out separately.
 const QUOTE_PAIRS: [(&str, &str); 3] = [("'", "'"), ("\"", "\""), ("`", "`")];
@@ -224,5 +224,95 @@ impl Editor {
                     })
                     .ok();
             }));
+    }
+}
+
+#[derive(Default, Debug)]
+pub struct SignatureHelpState {
+    task: Option<Task<()>>,
+    popover: Option<SignatureHelpPopover>,
+    hidden_by: Option<SignatureHelpHiddenBy>,
+    backspace_pressed: bool,
+}
+
+impl SignatureHelpState {
+    pub fn set_task(&mut self, task: Task<()>) {
+        self.task = Some(task);
+        self.hidden_by = None;
+    }
+
+    pub fn kill_task(&mut self) {
+        self.task = None;
+    }
+
+    #[cfg(test)]
+    pub fn popover(&self) -> Option<&SignatureHelpPopover> {
+        self.popover.as_ref()
+    }
+
+    pub fn popover_mut(&mut self) -> Option<&mut SignatureHelpPopover> {
+        self.popover.as_mut()
+    }
+
+    pub fn backspace_pressed(&self) -> bool {
+        self.backspace_pressed
+    }
+
+    pub fn set_backspace_pressed(&mut self, backspace_pressed: bool) {
+        self.backspace_pressed = backspace_pressed;
+    }
+
+    pub fn set_popover(&mut self, popover: SignatureHelpPopover) {
+        self.popover = Some(popover);
+        self.hidden_by = None;
+    }
+
+    pub fn hide(&mut self, hidden_by: SignatureHelpHiddenBy) {
+        if self.hidden_by.is_none() {
+            self.popover = None;
+            self.hidden_by = Some(hidden_by);
+        }
+    }
+
+    pub fn hidden_by_selection(&self) -> bool {
+        self.hidden_by == Some(SignatureHelpHiddenBy::Selection)
+    }
+
+    pub fn is_shown(&self) -> bool {
+        self.popover.is_some()
+    }
+}
+
+#[cfg(test)]
+impl SignatureHelpState {
+    pub fn task(&self) -> Option<&Task<()>> {
+        self.task.as_ref()
+    }
+}
+
+#[derive(Clone, Debug, PartialEq)]
+pub struct SignatureHelpPopover {
+    pub label: SharedString,
+    pub style: TextStyle,
+    pub highlights: Vec<(Range<usize>, HighlightStyle)>,
+}
+
+impl SignatureHelpPopover {
+    pub fn render(&mut self, max_size: Size<Pixels>, cx: &mut Context<Editor>) -> AnyElement {
+        div()
+            .id("signature_help_popover")
+            .elevation_2(cx)
+            .overflow_y_scroll()
+            .max_w(max_size.width)
+            .max_h(max_size.height)
+            .on_mouse_move(|_, _, cx| cx.stop_propagation())
+            .on_mouse_down(MouseButton::Left, |_, _, cx| cx.stop_propagation())
+            .child(
+                div().px_4().pb_1().child(
+                    StyledText::new(self.label.clone())
+                        .with_highlights(&self.style, self.highlights.iter().cloned()),
+                ),
+            )
+            .into_any_element()
     }
 }
