--- crates/editor/src/signature_help/state.rs
@@ -1,66 +1,0 @@
-use crate::signature_help::popover::SignatureHelpPopover;
-use crate::signature_help::SignatureHelpHiddenBy;
-use gpui::Task;
-
-#[derive(Default, Debug)]
-pub struct SignatureHelpState {
-    task: Option<Task<()>>,
-    popover: Option<SignatureHelpPopover>,
-    hidden_by: Option<SignatureHelpHiddenBy>,
-    backspace_pressed: bool,
-}
-
-impl SignatureHelpState {
-    pub fn set_task(&mut self, task: Task<()>) {
-        self.task = Some(task);
-        self.hidden_by = None;
-    }
-
-    pub fn kill_task(&mut self) {
-        self.task = None;
-    }
-
-    #[cfg(test)]
-    pub fn popover(&self) -> Option<&SignatureHelpPopover> {
-        self.popover.as_ref()
-    }
-
-    pub fn popover_mut(&mut self) -> Option<&mut SignatureHelpPopover> {
-        self.popover.as_mut()
-    }
-
-    pub fn backspace_pressed(&self) -> bool {
-        self.backspace_pressed
-    }
-
-    pub fn set_backspace_pressed(&mut self, backspace_pressed: bool) {
-        self.backspace_pressed = backspace_pressed;
-    }
-
-    pub fn set_popover(&mut self, popover: SignatureHelpPopover) {
-        self.popover = Some(popover);
-        self.hidden_by = None;
-    }
-
-    pub fn hide(&mut self, hidden_by: SignatureHelpHiddenBy) {
-        if self.hidden_by.is_none() {
-            self.popover = None;
-            self.hidden_by = Some(hidden_by);
-        }
-    }
-
-    pub fn hidden_by_selection(&self) -> bool {
-        self.hidden_by == Some(SignatureHelpHiddenBy::Selection)
-    }
-
-    pub fn is_shown(&self) -> bool {
-        self.popover.is_some()
-    }
-}
-
-#[cfg(test)]
-impl SignatureHelpState {
-    pub fn task(&self) -> Option<&Task<()>> {
-        self.task.as_ref()
-    }
-}
